Needed_items-menu without ajax
Based on the number of people in the market, the stock of clothes and the
estimated annual need of the products, this script estimates the need p.p. and the
effective need for a number of weeks. The number of weeks is input in a
first POST form. The calculation is then shown in a list.

- The menu only works for Camps with a market.
- The menu should be visible for all users.
- No AJAX is used.
- The _SESSION array has a new entry 'weeks'.