Create minified JS and CSS at 'build' time to avoid challenges in Google AppEngine
This process is a little flawed at the moment, given we have no hashes or cache busting in place, but can
still take the manual approach that was used before. Ultimately should be switching to webpack etc but
avoiding adopting too much tooling when we don't know the future lifetime of this code base yet.