Automatically deploy builds to Google App Engine.
When on the master branch
- deploy to staging environment (including running db migrations)
When on the production branch
- deploy to the demo environment (including running db migrations)
Intend to add production environment later.

We're using the 'services' feature of app engine to have these different
instances of the app, rather than 'versions'. Versions didn't allow us to have
seperate cron jobs for each environment, or other associated resources.