Introduce some really basic error reporting being sent to sentry.
I have rewritten the error throwing in database.php as it was unnecessarily wrapping & die-ing without giving more context than we'd get from the stack trace.

Removed the session scope as we're using it globally anyway and means it's safe to initialize twice.

Need to review all the random error level setting that goes on too, but trying to keep this minimal for now.