Consolidate error reporting so Sentry can report these, and we show (#18)
a error landing page even for code failures early in the loading of the
library stack

Lots of warnings are currently being supressed but can't easily resolve them,
so decided to preserve that behaviour.

Also now forcing the whole app to run through a single entry point again
to simplify where we hook up error handling. Hence including the htaccess necessary
for this.