Production merge 19/8/13 (#113)
* started with POST variable

* confirm message

* added line

* Removed Confirmation headline

* moved firstline into header

* add: shopping cart DOM elements to check_out

* add: initial shopping cart logic

* add: delete button in shopping list

* change: renderCart method, shopping cart sum

* add: change quantity in shopping cart possible, saving cart value in session storage

* add: update price per row instead of rerendering whole cart

* fix: cart styling and change quantity bugs, prevent cange quantity form submit

* add: submit shopping cart

* change: styling shopping cart changes

* change: product price stored per select option, no need to do extra backend request

* Fix form handler for shopping cart

* change: remove some costs fields, replace delete button with FA trash icon

* Add to cart unselects product

* Resolve shopping cart logic before submit

* Make Beneficiary comments searchable (#99)

* Add arabic and Kurmanji privacy texts

* Fix issue with products which have no value (#104)

* Copy guidelines from existing projects

* Add a missing global variable (#106)

They have to be declared here to work as we're wrapping the entry point in a function. I did another search for 'global $' in the code base to check if I missed any others, but hopefully got them all this time

* Replace any hard-coded HTTP urls to HTTPS

On the mobile login page, Cypress is causing errors when resources are loaded via HTTPS but then there's a HTTP redirect. Figured I might as well replace all occurrences rather than this one specific.

* Revert redirect to absolute but stay on HTTPS

We were redirecting with two // which broke the relative redirects, and given we have a potential open redirect vulnerability here figured it was safer to stick with absolute

* Another attempt to get the mobile redirect to stay on HTTPS

As seeing dropping down to HTTP in traces (and failing Cypress). Don't understand why.

* Ignore node_modules (introduced as part of cypress)

* Add cypress to cicleci build (against staging env)

Require the build step so we ensure packages are installed

* Add yarn and package.json so we can run cypress locally

* Cypress will pass CYPRESS_baseUrl env automatically

So we don't need to do something ourselves

* Run against staging, and so only after deploy

Ideally, we will run against each PR (in CI with it's own db in a container). But this is a good first step.

* Introduce a wait-on

Given the deploy has just happened we were getting timeouts

* Increase performance as we don't need to persist files

* Only run these tests on master after a deploy to staging

As we aren't currently spinning up MySQL as part of CI, we just feature test against the 'master' branch that is deployed on staging.

* Remove my name from author field