Add tracing as this page is very slow
In the hope we can track down where a big performance issue lies, add more tracing. To do this we have to wrap into trace functions, which then causes issues because of our implicit global variables - so we need to declare these.

I can't understand why $cmsmain can't be named in the use() list and instead has to be declared as a global.