Don't return success if the autologin fails
This caused a bug as we were saying the user is logged in when the session has timed out and we haven't been able to autologin. We really shouldn't be storing the md5 of the password client-side) nor should we be using md5, but unable to justify fixing this workflow now.