Escape all data passed to smarty templates by default
There are currently lots of potential XSS vulnerabilities - the latest I noticed was message strings being passed in the URL and served up as-is.

Switching this flag on will mean Smarty will encode all provided input to make it 'safe' for HTML output by default.

We will need to use 'nofilter' on anywhere that is 'correctly' unescaped.