Change Migration NewUserMgmt to rename settings table (#183)
I changed one of the first migrations to make it easier for the migration of Drop in the Ocean.

The settings table should be renamed from settings to cms_settings and not just dropped and created. The migration was created by the phinx generator and it did not recognize that a table was renamed.