Replace the N+1 db query with an in-memory approach
Instead of separately querying each row for its parent, we sort the results from the database
and order them such that for each parent, its direct children appear immediately next in the list.

Ideally we'd use a CTE for this, but Google Cloud SQL doesn't support MySQL 8 yet (which is when this became supported) so instead we'll do a poor man's version here and compute on the fly.

In order to preserve arbitrary sorting, we record the index of the row returned from the database, and
build a 'path' representing it.