Make loading AJAX actions more reliable in Cypress
The DropApp has strange behaviour where a button can trigger an AJAX call, but then subsequently trigger a reload of the same page. This messes up Cypress's reliablity because it can't tell when the DOM will reload. This change forces us to wait for the AJAX response to complete, and at least on my testing seems to be more reliable.